requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
replicate==0.8.3
Pillow==9.5.0
requests==2.31.0